requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-telegram-bot==20.7
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.3
python-dateutil==2.8.2

gunicorn==21.2.0
pywebpush==2.0.0
cryptography>=41.0.0
